[pandora car alarm] added channel "Mileage CAN"

+++ src/main/java/org/openhab/binding/pandoracaralarmsystem/internal/api/record/StatRecord.java
@@ -122,7 +122,7 @@ public class StatRecord {
      * The Mileage can.
      */
     @SerializedName("mileage_CAN")
-    public Float mileageCAN;
+    public Long mileageCAN;
 
     /**
      * The Metka.
